+++ src/navigation/AppNavigator.tsx
@@ -29,6 +29,7 @@ import { useChat } from "@context/ChatContext";
 import GoogleSignIn from "@screens/GoogleSignIn";
 import JoinRequestsScreen from "@screens/JoinRequestsScreen";
 import PendingRequestsScreen from "@screens/PendingRequestsScreen";
+import EditProfileScreen from "@screens/EditProfileScreen";
 import PastEventsScreen from "@screens/PastEventsScreen";
 import PrivacyPolicyScreen from "@screens/PrivacyPolicyScreen";
 import HelpScreen from "@screens/HelpScreen";
@@ -491,27 +492,35 @@ const AppNavigator = () => {
           }}
         />
         <Stack.Screen
+          name="EditProfile"
+          component={EditProfileScreen}
+          options={{
+            animation: "slide_from_right",
+            animationDuration: 350,
+          }}
+        />
+        <Stack.Screen
           name="PastEvents"
           component={PastEventsScreen}
           options={{
-            animation: "fade_from_bottom",
-            animationDuration: 200,
+            animation: "slide_from_right",
+            animationDuration: 350,
           }}
         />
         <Stack.Screen
           name="PrivacyPolicy"
           component={PrivacyPolicyScreen}
           options={{
-            animation: "fade_from_bottom",
-            animationDuration: 200,
+            animation: "slide_from_right",
+            animationDuration: 350,
           }}
         />
         <Stack.Screen
           name="Help"
           component={HelpScreen}
           options={{
-            animation: "fade_from_bottom",
-            animationDuration: 200,
+            animation: "slide_from_right",
+            animationDuration: 350,
           }}
         />
         <Stack.Screen
