+++ src/screens/ProfileScreen.tsx
@@ -1,5 +1,4 @@
 import {
-  ActivityIndicator,
   Alert,
   Animated,
   Image,
@@ -7,8 +6,6 @@ import {
   ScrollView,
   StyleSheet,
   Text,
-  TextInput,
-  TouchableOpacity,
   View,
 } from "react-native";
 
@@ -24,7 +21,7 @@ import { Feather } from "@expo/vector-icons";
 
 import ScreenContainer from "@components/ScreenContainer";
 import { colors, spacing, typography } from "@theme/index";
-import { useAuth, type ApiError } from "@context/AuthContext";
+import { useAuth } from "@context/AuthContext";
 import { useEvents } from "@context/EventsContext";
 import { useChat } from "@context/ChatContext";
 import { RootStackParamList, RootTabParamList } from "@navigation/types";
@@ -36,15 +33,8 @@ import PrivacyPolicyIcon from "@assets/account-icons/privacy.svg";
 import HelpIcon from "@assets/account-icons/help.svg";
 import LogoutIcon from "@assets/account-icons/logout.svg";
 import TrashIcon from "@assets/account-icons/delete.svg";
-import CameraIcon from "@assets/camera.svg";
 import * as Haptics from "expo-haptics";
 
-let ImagePicker: typeof import("expo-image-picker") | null = null;
-try {
-  ImagePicker = require("expo-image-picker");
-} catch (e) {
-  // expo-image-picker not available in Expo Go
-}
 
 type ProfileNavigation = CompositeNavigationProp<
   BottomTabNavigationProp<RootTabParamList, "Profile">,
@@ -84,7 +74,7 @@ const MenuItem = ({
 };
 
 const ProfileScreen = () => {
-  const { user, signOut, updateProfile } = useAuth();
+  const { user, signOut } = useAuth();
   const { events, userEvents } = useEvents();
   const { conversations } = useChat();
   const navigation = useNavigation<ProfileNavigation>();
@@ -125,107 +115,9 @@ const ProfileScreen = () => {
     signOut();
   }, [signOut]);
 
-  // Edit profile modal state
-  const [editProfileVisible, setEditProfileVisible] = useState(false);
-  const [editName, setEditName] = useState("");
-  const [editAvatarBase64, setEditAvatarBase64] = useState<string | null>(null);
-  const [removedAvatar, setRemovedAvatar] = useState(false);
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const editAvatarUri = removedAvatar
-    ? null
-    : editAvatarBase64
-      ? `data:image/jpeg;base64,${editAvatarBase64}`
-      : user?.avatar
-        ? `data:image/jpeg;base64,${user.avatar}`
-        : null;
-
-  const editHasChanges =
-    editName.trim() !== (user?.name ?? "") ||
-    editAvatarBase64 !== null ||
-    removedAvatar;
-
   const handleEditProfile = useCallback(() => {
-    setEditName(user?.name ?? "");
-    setEditAvatarBase64(null);
-    setRemovedAvatar(false);
-    setEditProfileVisible(true);
-  }, [user]);
-
-  const pickImage = useCallback(async () => {
-    void Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-    if (!ImagePicker) {
-      Alert.alert(
-        "Not Available",
-        "Image picker requires a native rebuild. Avatar upload is disabled.",
-      );
-      return;
-    }
-
-    const permissionResult =
-      await ImagePicker.requestMediaLibraryPermissionsAsync();
-    if (!permissionResult.granted) {
-      Alert.alert(
-        "Permission Required",
-        "Please allow access to your photo library to upload an avatar.",
-      );
-      return;
-    }
-
-    const result = await ImagePicker.launchImageLibraryAsync({
-      mediaTypes: ["images"],
-      allowsEditing: true,
-      aspect: [1, 1],
-      quality: 0.5,
-      base64: true,
-    });
-
-    if (!result.canceled && result.assets[0]?.base64) {
-      setEditAvatarBase64(result.assets[0].base64);
-      setRemovedAvatar(false);
-    }
-  }, []);
-
-  const handleRemoveAvatar = useCallback(() => {
-    void Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-    setEditAvatarBase64(null);
-    setRemovedAvatar(true);
-  }, []);
-
-  const handleSaveProfile = useCallback(async () => {
-    if (!editName.trim() || !user) return;
-
-    void Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
-    setIsSubmitting(true);
-    try {
-      let avatarValue: string | undefined;
-      if (editAvatarBase64) {
-        avatarValue = editAvatarBase64;
-      } else if (removedAvatar) {
-        avatarValue = undefined;
-      } else {
-        avatarValue = user.avatar ?? undefined;
-      }
-
-      await updateProfile({
-        name: editName.trim(),
-        gender: user.gender ?? "Male",
-        age: user.age ?? 18,
-        avatar: avatarValue,
-      });
-      setEditProfileVisible(false);
-    } catch (error) {
-      const status =
-        error instanceof Error ? (error as ApiError).status : undefined;
-      if (status === 401) return;
-      Alert.alert(
-        "Error",
-        error instanceof Error ? error.message : "Failed to save profile.",
-      );
-    } finally {
-      setIsSubmitting(false);
-    }
-  }, [editName, editAvatarBase64, removedAvatar, user, updateProfile]);
+    navigation.navigate("EditProfile");
+  }, [navigation]);
 
   const handlePastEvents = useCallback(() => {
     navigation.navigate("PastEvents");
@@ -389,91 +281,6 @@ const ProfileScreen = () => {
           />
         </View>
       </ScrollView>
-      <BottomSheetModal
-        visible={editProfileVisible}
-        onClose={() => setEditProfileVisible(false)}
-      >
-        {/* Header */}
-        <View style={styles.editHeader}>
-          <Text style={styles.editTitle}>Edit Profile</Text>
-          <Pressable
-            onPress={() => setEditProfileVisible(false)}
-            style={styles.editCloseButton}
-            accessibilityRole="button"
-          >
-            <Feather name="x" size={18} color="#999999" />
-          </Pressable>
-        </View>
-
-        {/* Avatar */}
-        <View style={styles.editAvatarSection}>
-          <View style={styles.editAvatarWrapper}>
-            <TouchableOpacity onPress={pickImage} accessibilityRole="button">
-              {editAvatarUri ? (
-                <Image
-                  source={{ uri: editAvatarUri }}
-                  style={styles.editAvatarImage}
-                />
-              ) : (
-                <LinearGradient
-                  colors={["#818CF8", "#6366F1", "#8B5CF6"]}
-                  start={{ x: 0, y: 0 }}
-                  end={{ x: 1, y: 1 }}
-                  style={styles.editAvatarGradient}
-                >
-                  <Text style={styles.editAvatarInitial}>
-                    {user?.name?.charAt(0).toUpperCase() ?? "?"}
-                  </Text>
-                </LinearGradient>
-              )}
-              <View style={styles.editCameraBadge}>
-                <CameraIcon width={14} height={14} />
-              </View>
-            </TouchableOpacity>
-            {editAvatarUri && (
-              <TouchableOpacity
-                style={styles.editRemoveBadge}
-                onPress={handleRemoveAvatar}
-                accessibilityRole="button"
-                accessibilityLabel="Remove photo"
-              >
-                <Feather name="x" size={13} color="#FFFFFF" />
-              </TouchableOpacity>
-            )}
-          </View>
-
-          {/* Name Input */}
-          <TextInput
-            style={styles.editNameInput}
-            value={editName}
-            onChangeText={setEditName}
-            placeholder="Your Name"
-            placeholderTextColor="#9CA3AF"
-            autoCapitalize="words"
-            autoCorrect={false}
-            textAlign="center"
-            returnKeyType="done"
-            maxLength={50}
-          />
-        </View>
-
-        {/* Save Button */}
-        <Pressable
-          style={[
-            styles.editSaveButton,
-            (!editHasChanges || isSubmitting) && styles.editSaveButtonDisabled,
-          ]}
-          onPress={handleSaveProfile}
-          disabled={!editHasChanges || isSubmitting}
-          accessibilityRole="button"
-        >
-          {isSubmitting ? (
-            <ActivityIndicator size="small" color="#FFFFFF" />
-          ) : (
-            <Text style={styles.editSaveButtonText}>Save</Text>
-          )}
-        </Pressable>
-      </BottomSheetModal>
     </ScreenContainer>
   );
 };
@@ -645,115 +452,6 @@ const styles = StyleSheet.create({
     fontSize: 16,
     fontFamily: typography.fontFamilySemiBold,
     color: "#FFFFFF",
-  },
-  // Edit Profile Modal
-  editHeader: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginBottom: spacing.sm,
-  },
-  editTitle: {
-    fontSize: 20,
-    fontFamily: typography.fontFamilySemiBold,
-    color: "rgba(0, 0, 0, 1)",
-    lineHeight: 24,
-    letterSpacing: -0.5,
-  },
-  editCloseButton: {
-    width: 32,
-    height: 32,
-    borderRadius: 80,
-    backgroundColor: "rgba(120, 120, 128, 0.16)",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  editAvatarSection: {
-    alignItems: "center",
-    marginBottom: spacing.md,
-  },
-  editAvatarWrapper: {
-    position: "relative",
-    marginBottom: 16,
-  },
-  editAvatarGradient: {
-    width: 80,
-    height: 80,
-    borderRadius: 80,
-    borderWidth: 2,
-    borderColor: "#E6E6E6",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  editAvatarImage: {
-    width: 80,
-    height: 80,
-    borderRadius: 80,
-    borderWidth: 2,
-    borderColor: "#E6E6E6",
-  },
-  editAvatarInitial: {
-    fontSize: 28,
-    color: "#FFFFFF",
-    fontFamily: typography.fontFamilyBold,
-  },
-  editCameraBadge: {
-    position: "absolute",
-    bottom: -2,
-    right: -4,
-    width: 28,
-    height: 28,
-    borderRadius: 14,
-    backgroundColor: "white",
-    justifyContent: "center",
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  editRemoveBadge: {
-    position: "absolute",
-    top: -2,
-    right: -2,
-    width: 22,
-    height: 22,
-    borderRadius: 11,
-    backgroundColor: "#000000",
-    justifyContent: "center",
-    alignItems: "center",
-    shadowColor: "#000",
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.15,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  editNameInput: {
-    fontFamily: typography.fontFamilyMedium,
-    fontSize: 24,
-    color: "#000000",
-    textAlign: "center",
-    paddingVertical: 8,
-    minWidth: 200,
-    letterSpacing: -0.5,
-  },
-  editSaveButton: {
-    backgroundColor: colors.primary,
-    borderRadius: 999,
-    paddingVertical: spacing.md,
-    alignItems: "center",
-  },
-  editSaveButtonDisabled: {
-    opacity: 0.5,
-  },
-  editSaveButtonText: {
-    color: colors.buttonText,
-    fontSize: 17,
-    fontFamily: typography.fontFamilyMedium,
-    lineHeight: 24,
-    letterSpacing: -0.5,
-    textAlign: "center",
   },
 });
